Separated some data processing steps out to data_processing.py, all models now automatically use exponential retries, examples have correct prompting strategy set to default. Everything works now.
