+++ .gitignore
@@ -20,8 +20,8 @@ factgenie/data/datasets.yml
 factgenie/data/inputs
 factgenie/data/outputs
 factgenie/config/config.yml
-factgenie/config/llm-eval
-factgenie/config/llm-gen
+# factgenie/config/llm-eval
+# factgenie/config/llm-gen
 factgenie/config/crowdsourcing
 
 # ignore changes in the notebook by default (to avoid adding the notebook after each run)
