+++ factgenie/app.py
@@ -633,11 +633,8 @@ def llm_campaign_new():
 
     # get a list of available metrics
     llm_configs = workflows.load_configs(mode=mode)
-    metric_types = list(ModelFactory.model_classes()[mode].keys())
-
-    # TEMP:
     model_apis = list(ModelFactory.get_model_apis().keys())
-    model_apis = [f"{api} API" for api in model_apis] 
+    prompt_strats = list(ModelFactory.get_prompt_strategies()[mode].keys())
 
     default_campaign_id = workflows.generate_default_id(app, mode=mode, prefix=mode.replace("_", "-"))
     default_prompts = utils.load_default_prompts()
@@ -650,8 +647,8 @@ def llm_campaign_new():
         default_prompts=default_prompts,
         available_data=available_data,
         configs=llm_configs,
-        metric_types=metric_types,
         model_apis=model_apis,
+        prompt_strats=prompt_strats,
         host_prefix=app.config["host_prefix"],
     )
