+++ factgenie/config/llm-eval/gemini-example.yaml
@@ -19,6 +19,7 @@ model: gemini-1.5-pro
 model_args:
   max_tokens: '1024'
   temperature: '0'
+prompt_strat: default
 prompt_template: |-
   Given the data:
   ```
