+++ factgenie/config/llm-gen/anthropic-example.yaml
@@ -3,6 +3,7 @@ extra_args:
   remove_suffix: '"'
 model: claude-3-5-haiku-20241022
 model_args: {}
+prompt_strat: default
 prompt_template: |-
   Given the data:
   ```
