+++ factgenie/config/llm-gen/gemini-example.yaml
@@ -3,6 +3,7 @@ extra_args:
   remove_suffix: '"'
 model: gemini-1.5-pro
 model_args: {}
+prompt_strat: default
 prompt_template: |-
   Given the data:
   ```
