+++ factgenie/config/llm-gen/openai-example.yaml
@@ -6,6 +6,7 @@ system_msg: "You are an expert data-to-text system. You understand structured da
 start_with: "Sure, here is the required output:\n\""
 extra_args:
   remove_suffix: '"'
+prompt_strat: default
 prompt_template: |
   Given the data:
   ```
