+++ factgenie/llm_campaign.py
@@ -264,7 +264,8 @@ def parse_llm_gen_config(config):
 
 def parse_llm_eval_config(config):
     config = {
-        "type": config.get("metricType"),
+        "type": config.get("metricType"), # metric type is now api type
+        "prompt_type": config.get("promptStrat"),
         "model": config.get("modelName"),
         "prompt_template": config.get("promptTemplate"),
         "system_msg": config.get("systemMessage"),
