+++ factgenie/static/js/campaigns.js
@@ -222,7 +222,8 @@ function gatherConfig() {
         config.sliders = getSliders();
         config.textFields = getKeys($("#textFields"));
     } else if (window.mode == "llm_eval" || window.mode == "llm_gen") {
-        config.metricType = $("#metric-type").val();
+        config.metricType = $("#model-api").val(); // metric type is now api type
+        config.promptStrat = $("#prompt-strat").val();
         config.modelName = $("#model-name").val();
         config.modelApi = $("#model-api").val();
         config.promptTemplate = $("#prompt-template").val();
@@ -625,7 +626,8 @@ function updateLLMMetricConfig() {
     }
     const cfg = window.configs[llmConfigValue];
 
-    const metric_type = cfg.type;
+    const api_type = cfg.type;
+    const prompt_strat = cfg.prompt_strat;
     const model_name = cfg.model;
     const prompt_template = cfg.prompt_template;
     const system_msg = cfg.system_msg;
@@ -635,7 +637,8 @@ function updateLLMMetricConfig() {
     const extra_args = cfg.extra_args;
 
     // for metric, we need to select the appropriate one from the values in the select box
-    $("#metric-type").val(metric_type);
+    $("#model-api").val(api_type);
+    $("#prompt-strat").val(prompt_strat);
     $("#model-name").html(model_name);
     $("#prompt-template").html(prompt_template);
     $("#system-message").html(system_msg);
