+++ factgenie/templates/pages/llm_campaign_new.html
@@ -91,6 +91,24 @@
 
               <div class="row g-3 mb-3 align-items-center">
                 <div class="col-auto">
+                  <i class="fa fa-comments"></i>
+                  <span style="margin-left: 5px; margin-right: 10px;">Prompting strategy</span>
+                  <div class="mb-2">
+                    <small class="form-text text-muted">Select one of the available prompting strategies (or add your own class to <code>factgenie/models.py</code>). A prompting strategy controls how the prompt will be used.
+                    </small>
+                  </div>
+                </div>
+                <div class="form-group mt-1">
+                  <select class="form-select" id="prompt-strat" name="prompt-strat">
+                    {% for promt_strat in prompt_strats %}
+                    <option value="{{ promt_strat }}">{{ promt_strat }}</option>
+                    {% endfor %}
+                  </select>
+                </div>
+              </div>
+
+              <div class="row g-3 mb-3 align-items-center">
+                <div class="col-auto">
                   <i class="fa fa-align-left"></i>
                   <span style="margin-left: 5px; margin-right: 10px;">Prompt template</span>
                   <div class="mb-2">
@@ -140,35 +158,14 @@
               </div>
               {% endif %}
 
-              <hr style="margin-left: -30px; margin-right: -30px">
+              <hr style="margin-left: -30px; margin-right: -30px; margin-top: 30px; margin-bottom: 25px;">
 
               <div class="row g-3 mb-3 align-items-center">
                 <div class="col-auto">
                   <i class="fa fa-code"></i>
                   <span style="margin-left: 5px; margin-right: 10px;">API provider</span>
                   <div class="mb-2">
-                    <small class="form-text text-muted">Select one of the available providers (or add your own
-                      class
-                      to
-                      <code>factgenie/models.py</code>) Do not forget to add the relevant API keys in the <a href="/app_config">app settings</a> or in the <code>factgenie/config/config.yml</code> file.
-                    </small>
-                  </div>
-                </div>
-                <div class="form-group mt-1">
-                  <select class="form-select" id="metric-type" name="metric-type">
-                    {% for metric_type in metric_types %}
-                    <option value="{{ metric_type }}">{{ metric_type }}</option>
-                    {% endfor %}
-                  </select>
-                </div>
-              </div>
-
-              <div class="row g-3 mb-3 align-items-center">
-                <div class="col-auto">
-                  <i class="fa fa-code"></i>
-                  <span style="margin-left: 5px; margin-right: 10px;">API provider</span>
-                  <div class="mb-2">
-                    <small class="form-text text-muted">Select one of the available providers (or add your own class to <code>factgenie/models.py</code>) Do not forget to add the relevant API keys in the <a href="/app_config">app settings</a> or in the <code>factgenie/config/config.yml</code> file.
+                    <small class="form-text text-muted">Select one of the available providers (or add your own class to <code>factgenie/models.py</code>). Do not forget to add the relevant API keys in the <a href="/app_config">app settings</a> or in the <code>factgenie/config/config.yml</code> file.
                     </small>
                   </div>
                 </div>
